requirements.txt
# Core dependencies
pywebview
requests
flask
flask-cors
packaging

# Printer drivers
python-escpos  # Universal ESC/POS (Star, Epson, Citizen, Generic)
bleak          # Bluetooth Low Energy scanning
pyserial       # Serial port communication
pyusb          # USB communication
pycups         # CUPS integration (macOS/Linux)

# Audio & Media
pygame

# Build
pyinstaller
